requirements.txt
streamlit
requests
pillow
python-dotenv
fpdf
